`update-alternatives` with version specific target
The `python-rpm-macros` will look for all `python_flavors` in
`/usr/bin`, this means if a user specifies `python3.10` as a flavor then
their RPM build looks for `/usr/bin/python3.10`. In order to support
this, we must make version specific updates to `/usr/bin` as well as the
major version updates.