Remove compile step and build for one distro
Removing the compiling step will make this build a lot faster, and since
we no longer care about granular control over the Python version we
might as well just use the OS package management for it.

This simplifies the build, removing the matrix, and making this build
responsible only for the latest Python version available in the distro.
SLES15SP4 has Python3.6 and Python3.10 available, this build will now
build Python3.10 and SP4 only.

If this commit is backported to a `maint/` branch, then that can build an
older distro and the older Python accordingly (e.g. `maint/3.9` should be
based in SLES15SP3, and use `pythonVersion = 3.9`).